// frontend/src/api/sdk/api.ts
export interface Job {
    'job_id': string;
    'filename': string;
    'status': string;
    'transcript'?: string | null;
    'owner': string;
    'stored_filename': string;
    'error_message'?: string | null;
}